+++ src-docs/src/views/combo_box/colors.js
@@ -90,7 +90,7 @@ export default () => {
     /* DisplayToggles wrapper for Docs only */
     <DisplayToggles canDisabled={false} canReadOnly={false} canIsDisabled>
       <EuiComboBox
-        ariaLabel="Accessible screen reader label"
+        aria-label="Accessible screen reader label"
         placeholder="Select or create options"
         options={options}
         selectedOptions={selectedOptions}