+++ src-docs/src/views/combo_box/combo_box.js
@@ -74,7 +74,7 @@ export default () => {
     /* DisplayToggles wrapper for Docs only */
     <DisplayToggles canDisabled={false} canReadOnly={false} canIsDisabled>
       <EuiComboBox
-        ariaLabel="Accessible screen reader label"
+        aria-label="Accessible screen reader label"
         placeholder="Select or create options for"
         options={options}
         selectedOptions={selectedOptions}