+++ src-docs/src/views/combo_box/combo_box_example.js
@@ -19,7 +19,7 @@ import ComboBox from './combo_box';
 const comboBoxSource = require('!!raw-loader!./combo_box');
 const comboBoxHtml = renderToHtml(ComboBox);
 const comboBoxSnippet = `<EuiComboBox
-  ariaLabel="Accessible screen reader label"
+  aria-label="Accessible screen reader label"
   placeholder="Select or create options"
   options={[
     {
@@ -39,7 +39,7 @@ import Colors from './colors';
 const colorsSource = require('!!raw-loader!./colors');
 const colorsHtml = renderToHtml(Colors);
 const colorsSnippet = `<EuiComboBox
-  ariaLabel="Accessible screen reader label"
+  aria-label="Accessible screen reader label"
   placeholder="Select or create options"
   options={[
     {
@@ -57,7 +57,7 @@ import RenderOption from './render_option';
 const renderOptionSource = require('!!raw-loader!./render_option');
 const renderOptionHtml = renderToHtml(RenderOption);
 const renderOptionSnippet = `<EuiComboBox
-  ariaLabel="Accessible screen reader label"
+  aria-label="Accessible screen reader label"
   placeholder="Select or create options"
   options={options}
   selectedOptions={selectedOptions}
@@ -70,7 +70,7 @@ import Groups from './groups';
 const groupsSource = require('!!raw-loader!./groups');
 const groupsHtml = renderToHtml(Groups);
 const groupsSnippet = `<EuiComboBox
-  ariaLabel="Accessible screen reader label"
+  aria-label="Accessible screen reader label"
   placeholder="These options are grouped"
   options={[colorGroup, soundGroup]}
   selectedOptions={selectedOptions}
@@ -84,7 +84,7 @@ const singleSelectionSource = require('!!raw-loader!./single_selection');
 const singleSelectionHtml = renderToHtml(SingleSelection);
 const singleSelectionPrependSource = require('!!raw-loader!./single_selection_prepend');
 const singleSelectionSnippet = `<EuiComboBox
-  ariaLabel="Accessible screen reader label"
+  aria-label="Accessible screen reader label"
   placeholder="Select a single option"
   singleSelection={{ asPlainText: true }}
   options={options}
@@ -105,7 +105,7 @@ const singleSelectionCustomOptionsHtml = renderToHtml(
   SingleSelectionCustomOptions
 );
 const singleSelectionCustomOptionsSnippet = `<EuiComboBox
-  ariaLabel="Accessible screen reader label"
+  aria-label="Accessible screen reader label"
   placeholder="Select a single option"
   singleSelection={{ asPlainText: true }}
   options={options}
@@ -118,7 +118,7 @@ import DisallowCustomOptions from './disallow_custom_options';
 const disallowCustomOptionsSource = require('!!raw-loader!./disallow_custom_options');
 const disallowCustomOptionsHtml = renderToHtml(DisallowCustomOptions);
 const disallowCustomOptionsSnippet = `<EuiComboBox
-  ariaLabel="Accessible screen reader label"
+  aria-label="Accessible screen reader label"
   placeholder="Select one or more options"
   options={options}
   onChange={onChange}
@@ -144,7 +144,7 @@ const asyncHtml = renderToHtml(Async);
 const asyncSnippet = `<EuiComboBox
   placeholder="Search asynchronously"
   async
-  ariaLabel="Accessible screen reader label"
+  aria-label="Accessible screen reader label"
   options={options}
   selectedOptions={selectedOptions}
   isLoading={isLoading}
@@ -157,7 +157,7 @@ import Virtualized from './virtualized';
 const virtualizedSource = require('!!raw-loader!./virtualized');
 const virtualizedHtml = renderToHtml(Virtualized);
 const virtualizedSnippet = `<EuiComboBox
-  ariaLabel="Accessible screen reader label"
+  aria-label="Accessible screen reader label"
   placeholder="Select one or more options"
   options={options}
   selectedOptions={selectedOptions}
@@ -168,7 +168,7 @@ import Disabled from './disabled';
 const disabledSource = require('!!raw-loader!./disabled');
 const disabledHtml = renderToHtml(Disabled);
 const disabledSnippet = `<EuiComboBox
-  ariaLabel="Accessible screen reader label"
+  aria-label="Accessible screen reader label"
   placeholder="Select or create options"
   options={options}
   onChange={onChange}
@@ -180,7 +180,7 @@ import Delimiter from './combo_box_delimiter';
 const delimiterSource = require('!!raw-loader!./combo_box_delimiter');
 const delimiterHtml = renderToHtml(Delimiter);
 const delimiterSnippet = `<EuiComboBox
-  ariaLabel="Accessible screen reader label"
+  aria-label="Accessible screen reader label"
   placeholder="Select or create options"
   options={options}
   delimiter=","
@@ -194,7 +194,7 @@ import StartingWith from './startingWith';
 const startingWithSource = require('!!raw-loader!./startingWith');
 const startingWithHtml = renderToHtml(StartingWith);
 const startingWithSnippet = `<EuiComboBox
-  ariaLabel="Accessible screen reader label"
+  aria-label="Accessible screen reader label"
   placeholder="Select or create options"
   sortMatchesBy="startsWith"
   options={options}
@@ -220,7 +220,7 @@ import Labelledby from './combo_box_labelledby';
 const labelledbySource = require('!!raw-loader!./combo_box_labelledby');
 const labelledbyHtml = renderToHtml(DuplicateOptions);
 const labelledbySnippet = `<EuiComboBox
-  ariaLabelledby={generatedId}
+  aria-labelledby={generatedId}
   placeholder="Select or create options"
   options={options}
   selectedOptions={selectedOptions}
@@ -245,8 +245,8 @@ export const ComboBoxExample = {
           You must add an accessible label to each instance of{' '}
           <strong>EuiComboBox</strong>. Labels can be created by wrapping the
           combobox in an <strong>EuiFormRow</strong>, adding an{' '}
-          <EuiCode language="js">ariaLabel</EuiCode> prop, or passing a text
-          node ID to the <EuiCode language="js">ariaLabelledby</EuiCode> prop.
+          <EuiCode language="js">aria-label</EuiCode> prop, or passing a text
+          node ID to the <EuiCode language="js">aria-labelledby</EuiCode> prop.
         </p>
       </EuiText>
     </Fragment>
@@ -668,7 +668,7 @@ export const ComboBoxExample = {
           paragraph. You can easily create a unique ID for a text element using
           the <a href="/#/utilities/html-id-generator">HTML ID generator</a>.
           Then pass your unique ID to the{' '}
-          <EuiCode language="js">ariaLabelledby</EuiCode> prop.
+          <EuiCode language="js">aria-labelledby</EuiCode> prop.
         </p>
       ),
       props: { EuiComboBox, EuiComboBoxOptionOption },