+++ src-docs/src/views/combo_box/virtualized.js
@@ -24,7 +24,7 @@ export default () => {
 
   return (
     <EuiComboBox
-      ariaLabel="Accessible screen reader label"
+      aria-label="Accessible screen reader label"
       placeholder="Select one or more options"
       options={options}
       selectedOptions={selectedOptions}